refactor(stages): move Spec accessors to MetricSpec methods; split export test
