+++ internal/stages/metrics.go
@@ -7,36 +7,16 @@ import (
 	"github.com/theunrepentantgeek/code-visualizer/internal/provider"
 )
 
-// SpecMetric returns the metric name from a *MetricSpec, or "" if nil.
-func SpecMetric(s *config.MetricSpec) metric.Name {
-	if s == nil {
-		return ""
-	}
-
-	return s.Metric
-}
-
-// SpecPalette returns the palette name from a *MetricSpec, or "" if nil.
-func SpecPalette(s *config.MetricSpec) palette.PaletteName {
-	if s == nil {
-		return ""
-	}
-
-	return s.Palette
-}
-
 // CollectRequestedMetrics returns the unique ordered list of metric names
 // implied by size + optional fill + optional border specs.
-func CollectRequestedMetrics(size metric.Name, fill, border *config.MetricSpec) []metric.Name {
+func CollectRequestedMetrics(size metric.Name, fill *config.MetricSpec, border *config.MetricSpec) []metric.Name {
 	seen := map[metric.Name]bool{size: true}
 	names := []metric.Name{size}
 
 	for _, spec := range []*config.MetricSpec{fill, border} {
-		if spec != nil && spec.Metric != "" {
-			if !seen[spec.Metric] {
-				seen[spec.Metric] = true
-				names = append(names, spec.Metric)
-			}
+		if name := spec.MetricName(); name != "" && !seen[name] {
+			seen[name] = true
+			names = append(names, name)
 		}
 	}
 
@@ -46,7 +26,7 @@ func CollectRequestedMetrics(size metric.Name, fill, border *config.MetricSpec) 
 // ResolveFillPalette returns the fill palette to use, consulting (in order)
 // the explicit fill spec, the provider's default palette, and palette.Neutral.
 func ResolveFillPalette(fill *config.MetricSpec, fillMetric metric.Name) palette.PaletteName {
-	if fp := SpecPalette(fill); fp != "" {
+	if fp := fill.PaletteName(); fp != "" {
 		return fp
 	}
 
@@ -60,12 +40,12 @@ func ResolveFillPalette(fill *config.MetricSpec, fillMetric metric.Name) palette
 // ResolveBorderMetricAndPalette returns the effective border metric and
 // palette name, or ("", "") when no border is configured.
 func ResolveBorderMetricAndPalette(border *config.MetricSpec) (metric.Name, palette.PaletteName) {
-	borderMetric := SpecMetric(border)
+	borderMetric := border.MetricName()
 	if borderMetric == "" {
 		return "", ""
 	}
 
-	borderPaletteName := SpecPalette(border)
+	borderPaletteName := border.PaletteName()
 	if borderPaletteName == "" {
 		if d, ok := provider.GetDescriptor(borderMetric); ok {
 			borderPaletteName = d.DefaultPalette
