+++ internal/stages/scan_test.go
@@ -11,18 +11,6 @@ import (
 	"github.com/theunrepentantgeek/code-visualizer/internal/stages"
 )
 
-func TestExportConfig_NoFlag_NoOp(t *testing.T) {
-	t.Parallel()
-	g := NewGomegaWithT(t)
-
-	s := &fakeState{common: stages.CommonState{
-		Flags:      &stages.Flags{ExportConfig: ""},
-		RootConfig: config.New(),
-	}}
-
-	g.Expect(stages.ExportConfig[*fakeState](s)).To(Succeed())
-}
-
 func TestResolveDimensions_AppliesDefaults(t *testing.T) {
 	t.Parallel()
 	g := NewGomegaWithT(t)
